refactor: merge DatasetServer into HistoryServer
- Delete standalone dataset module (handlers.rs, mod.rs, tools.rs)
- Move exportDataset tool into HistoryServer (tools.rs + handlers.rs)
- Add 'dataset' alias to History BuiltinServiceId for backward compat
- Remove DatasetServer registration from BuiltinServerRegistry (global + session)
- Remove Dataset factory from MCPServiceProxy
- Remove Dataset metadata/tools from server tool listing
- Update generated builtin-services.ts
