+++ src-tauri/src/commands/mod.rs
@@ -2,6 +2,7 @@ pub mod agent_commands;
 pub mod assistant_crud_commands;
 pub mod attachments_commands;
 pub mod browser_commands;
+pub mod dataset_commands;
 pub mod download_commands;
 pub mod file_commands;
 pub mod knowledge_commands;
@@ -17,4 +18,3 @@ pub mod skill_commands;
 pub mod skill_management;
 pub mod url_commands;
 pub mod workspace_commands;
-pub mod dataset_commands;
