+++ src-tauri/src/mcp/builtin/history/handlers.rs
@@ -810,3 +810,34 @@ fn render_search_text(page: &Page<HistorySearchMatch>, caller_session_id: &str) 
 
     lines.join("\n")
 }
+
+pub async fn export_dataset(args: Value) -> Result<MCPResult, String> {
+    use crate::commands::dataset_commands::{
+        export_dataset as run_export_dataset, DatasetFilter, ExportFormat,
+    };
+
+    let session_ids: Option<Vec<String>> = args
+        .get("sessionIds")
+        .and_then(|v| serde_json::from_value(v.clone()).ok());
+
+    let format: ExportFormat = args
+        .get("format")
+        .and_then(|v| serde_json::from_value(v.clone()).ok())
+        .ok_or_else(|| "Missing or invalid 'format' parameter".to_string())?;
+
+    let output_path: String = args
+        .get("outputPath")
+        .and_then(|v| v.as_str().map(|s| s.to_string()))
+        .ok_or_else(|| "Missing or invalid 'outputPath' parameter".to_string())?;
+
+    let filters: Option<DatasetFilter> = args
+        .get("filters")
+        .and_then(|v| serde_json::from_value(v.clone()).ok());
+
+    let result = run_export_dataset(session_ids, format, output_path, filters).await?;
+
+    Ok(MCPResult::success(&format!(
+        "Successfully exported dataset: {} sessions, {} messages written to {}",
+        result.session_count, result.message_count, result.output_path
+    )))
+}
