+++ src-tauri/src/mcp/service_proxy/factory.rs
@@ -104,8 +104,5 @@ pub(crate) async fn create_builtin_server(
         BuiltinServiceId::Media => Ok(Some(Box::new(
             crate::mcp::builtin::media::MediaServer::new(_session_id, _session_manager),
         ))),
-        BuiltinServiceId::Dataset => Ok(Some(Box::new(
-            crate::mcp::builtin::dataset::DatasetServer::new(),
-        ))),
     }
 }
